add header slider bg, fix process
