+++ .gitignore
@@ -2,3 +2,4 @@
 wp-config.php
 node_modules
 npm-debug.log
+wp-content/languages
