+++ wp-content/themes/giccanada/style.css
@@ -84,8 +84,7 @@ ol.carousel-indicators li.active {
 /*fixed right panel end -----------------------*/
 .header-container {
   width: 100%;
-  background: url(./public/images/header-bg-60bd5054a3.png) no-repeat;
-  background-size: cover; }
+  height: 828px; }
 
 /*top bar*/
 .top-bar {
@@ -177,7 +176,11 @@ ol.carousel-indicators li.active {
 /*top bar end --------------------*/
 /*menu*/
 .menu-container {
-  margin-top: 24px; }
+  margin-top: 24px;
+  position: absolute;
+  left: 50%;
+  margin-left: -570px;
+  z-index: 1; }
 
 .menu-logo {
   background: url(./public/images/Logo-f6c4044548.png) no-repeat;
@@ -241,8 +244,11 @@ ol.carousel-indicators li.active {
 
 /*menu end ------------------------------*/
 /*menu slider*/
-#header-carousel {
-  margin-top: 108px; }
+.carousel-div {
+  padding: 0; }
+  .carousel-div #header-carousel .carousel-item {
+    padding-top: 200px;
+    padding-bottom: 180px; }
 
 .hs-item {
   width: 570px; }
@@ -271,7 +277,19 @@ ol.carousel-indicators li.active {
 
 .hs-indicators {
   position: relative;
-  margin-top: 60px; }
+  top: -60px; }
+
+#hs-sl1 {
+  background: url(./public/images/Hero-image_1-5e56da563d.jpg) no-repeat;
+  background-size: 100%; }
+
+#hs-sl2 {
+  background: url(./public/images/Hero-image_2-ef6bff7fbf.jpg) no-repeat;
+  background-size: cover; }
+
+#hs-sl3 {
+  background: url(./public/images/Hero-image_3-c171fa890e.jpg) no-repeat;
+  background-size: cover; }
 
 /*menu slider end ------------------------------*/
 /*programms*/
@@ -463,7 +481,6 @@ ol.carousel-indicators li.active {
   color: #ff9500; }
 
 .process-text, .process-descr {
-  width: 570px;
   line-height: 24px;
   font-size: 1.25em;
   text-align: center;
@@ -471,7 +488,8 @@ ol.carousel-indicators li.active {
 
 .process-text {
   margin-top: 10px;
-  margin-bottom: 60px; }
+  margin-bottom: 60px;
+  width: 570px; }
 
 .process-spider-top {
   max-width: 400px; }
@@ -482,7 +500,7 @@ ol.carousel-indicators li.active {
 
 .process-descr {
   margin-top: 40px;
-  padding: 0 7%; }
+  padding: 0 18px; }
 
 .process-spider-bottom {
   max-width: 300px; }
@@ -757,6 +775,8 @@ ol.carousel-indicators li.active {
 /* lg-размер (<=1199px) */
 @media (max-width: 1199px) {
   /* CSS для ширины от 992px до 1199px */
+  .menu-container {
+    margin-left: -480px; }
   .top-bar-menu-link {
     font-size: .9em; }
   .tb-search {
@@ -804,7 +824,8 @@ ol.carousel-indicators li.active {
 @media (max-width: 991px) {
   /* CSS для ширины от 768px до 991px */
   .menu-container {
-    min-width: 780px; }
+    min-width: 780px;
+    margin-left: -390px; }
   .top-bar-menu-link {
     font-size: .8em; }
   .menu-item .menu-dropdown .dropdown-content.show-dropdown-content {
