+++ wp-content/themes/giccanada/template-parts/process.php
@@ -1,9 +1,9 @@
 <?php ?>
 
-<div class="row justify-content-center">
+<div class="row justify-content-center no-gutters">
 	<h2 class="col process-caption">Процесс</h2>
 </div>
-<div class="row justify-content-center">
+<div class="row justify-content-center no-gutters">
         <span class="center-block process-text">GIC Canada Immigration оформляет иммиграционные файлы своих клиентов только после
             тщательного изучения дела, объективной оценки шансов и подбора оптимальных вариантов.</span>
 </div>
@@ -37,7 +37,7 @@
 		</div>
 	</div>
 </div>
-<div class="row justify-content-center">
+<div class="row justify-content-center no-gutters">
 	<div class="process-spider"></div>
 </div>
 <div class="row justify-content-center no-gutters flex-nowrap">
@@ -76,6 +76,6 @@
 		</div>
 	</div>
 </div>
-<div class="row justify-content-center">
+<div class="row justify-content-center no-gutters">
 	<a href="#" class="col col-md-auto academy-slr-item-btn orange-btn">Заполнить анкету</a>
 </div>
